+++ include/pdfixsdksamples/OcrWithTesseract.h
@@ -11,6 +11,6 @@ void OcrWithTesseract(
     const std::wstring& save_path,                  // searchable PDF document
     const std::wstring& data_path,                  // path to OCR data
     const std::wstring& language,                   // default OCR language
-    const double zoom,                              // page zoom level for rendering to control image processing quality
+    const float zoom,                               // page zoom level for rendering to control image processing quality
     const PdfRotate rotate                          // page rotation
     );
